Move feedAlices to globalSetup.ts
